[fix] refactor for being able to load plugins
