+++ as2_behaviors/as2_behaviors_path_planning/plugins/a_star/CMakeLists.txt
@@ -1,0 +1,69 @@
+cmake_minimum_required(VERSION 3.5)
+set(PROJECT_NAME a_star)
+project(${PROJECT_NAME} VERSION 1.0.0)
+
+# Default to C++17
+if(NOT CMAKE_CXX_STANDARD)
+  set(CMAKE_CXX_STANDARD 17)
+endif()
+
+# set Release as default
+if(NOT CMAKE_BUILD_TYPE)
+  set(CMAKE_BUILD_TYPE Release)
+endif()
+
+# set fPIC to ON by default
+set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+
+# find dependencies
+set(PLUGIN_DEPENDENCIES
+  ament_cmake
+  rclcpp
+  pluginlib
+  as2_core
+  as2_msgs
+  geometry_msgs
+  Eigen3
+)
+
+foreach(DEPENDENCY ${PLUGIN_DEPENDENCIES})
+  find_package(${DEPENDENCY} REQUIRED)
+endforeach()
+
+include_directories(
+  include
+  include/${PROJECT_NAME}
+)
+
+set(SOURCE_CPP_FILES
+  src/${PROJECT_NAME}.cpp
+)
+
+# Library
+add_library(${PROJECT_NAME} SHARED ${SOURCE_CPP_FILES})
+target_link_libraries(${PROJECT_NAME} as2_behaviors_path_planning_plugin_base)
+ament_target_dependencies(${PROJECT_NAME} ${PLUGIN_DEPENDENCIES})
+
+install(
+  DIRECTORY include/
+  DESTINATION include
+)
+
+ament_export_include_directories(
+  include
+)
+ament_export_libraries(
+  ${PROJECT_NAME}
+)
+ament_export_targets(
+  export_${PROJECT_NAME}
+)
+
+install(
+  TARGETS ${PROJECT_NAME}
+  EXPORT export_${PROJECT_NAME}
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
+)
+
